+++ modules/default.nix
@@ -29,6 +29,11 @@ with lib;
       '';
       type = types.lines;
     };
+    keychainUnlock = mkOption {
+      default = true;
+      description = "Try unlock keychain on SSH";
+      type = types.bool;
+    };
     ohMyZsh.enable = mkOption {
       default = true;
       description = "oh-my-zsh integration";
@@ -111,6 +116,21 @@ with lib;
                   command nix "$@"
                 fi
               }
+            ''
+          ]
+          ++ optionals cfg.keychainUnlock [
+            ''
+              if [ -n "$SSH_CLIENT" ]; then
+                LOCKED="$(busctl --user get-property org.freedesktop.secrets /org/freedesktop/secrets/collection/login org.freedesktop.Secret.Collection Locked 2>/dev/null | awk '{printf "%s", $2}')"
+
+                if "$LOCKED"; then
+                  printf 'Unlock keyring (enter password for %s): ' "$USER"
+                  read -s PASS
+                  echo ""
+                  printf '%s' "$PASS" | gnome-keyring-daemon --unlock --replace >/dev/null 2>&1
+                  unset PASS
+                fi
+              fi
             ''
           ]
           ++ [ cfg.initContent ]